requirements/dev.requirements.txt
alembic
asyncpg
email-validator
fastapi
pydantic-settings
SQLAlchemy
uvicorn[standard]
